test: raise line+branch coverage to 100% and remediate SonarCloud smells [full ci]
Coverage: add targeted host tests for reachable edges (ad9238 bad reg_addr, pcap/http_range bounds, clock zero-sample avg, arena set-free routing, crc out-of-range width clamp, h3 dispatch field mapping, ws permessage-deflate overflow, ssh_server message-switch arms, qpack/hpack truncated varints, numparse exponent/sign, trace_capture guards, tcp obs hook, udp freed-slot scan, http_parser tail flush, keymat/aesgcm counter carry). Annotate provably-dead defensive branches with justified GCOVR_EXCL.

Remediate 29 code smells at source: const-correct params (dbm/docstore), enum NOSONAR wire-constant tables, nested-ternary->clamp_width helper (crc), for->while loop restructures (scpi/snmp), merged ifs (webdav/run_tests), redundant-cast removal, nested-break flatten (oidc), cognitive-complexity extraction (quic_conn/hpack), empty-method comment (client), shell stderr redirects + if-merge (run_tests.sh).

+++ src/services/rcwl0516/rcwl0516.cpp
@@ -22,7 +22,7 @@ namespace
 // A limit of 0 is always satisfied, which is what disables debounce / hold.
 inline bool elapsed(uint32_t now, uint32_t since, uint32_t limit)
 {
-    return (uint32_t)(now - since) >= limit;
+    return (now - since) >= limit;
 }
 } // namespace
